[bugFix] fix multi_judger, ray dependency and sglang launch method
