+++ xtuner/v1/ray/judger/controller.py
@@ -96,6 +96,28 @@ class JudgerController:
         )
         self.logger = get_logger(log_dir=judger_config.worker_log_dir, tag="Judger")
 
+    async def _call_single_reward_judger(self, judger: NativeJudger, group_data_item: List[RLDataFlowItem]):
+        """Call a single custom reward judger to calculate rewards.
+
+        Args:
+            judger (NativeJudger): An instance of a custom judger.
+            responses (List[str]): A list of model-generated responses.
+            labels (List[str]): A list of ground-truth labels.
+
+        Returns:
+            List[RLJudgerResponseItem]: A list of RLJudgerResponseItem containing
+                calculated rewards for each sample.
+        """
+        tasks = []
+        if self.judger_config.enable_batch_reward:
+            task = judger.judge(group_data_item)
+            tasks.append(task)
+        else:
+            for data_item in group_data_item:
+                task = judger.judge([data_item])
+                tasks.append(task)
+        return tasks
+
     async def _call_custom_reward_judger(
         self,
         active_judgers: List[NativeJudger],
@@ -113,39 +135,36 @@ class JudgerController:
             Dict[str, List[float]]: A dictionary where keys are judger names
                 and values are lists of calculated rewards for each sample.
         """
-        results = []
-        if self.judger_config.enable_batch_reward:
-            tasks = [judger.judge(group_data_item) for judger in active_judgers]
-            results = await asyncio.gather(*tasks)
-        else:
-            tasks_per_sample = [
-                [(judger.judge([group_data_item[i]])) for judger in active_judgers]
-                for i in range(len(group_data_item))
-            ]
+        active_judgers_len = len(active_judgers)
+        task_len_list = [0]
+        judger_list = []
+        all_tasks = []
+        for judger in active_judgers:
+            tasks = await self._call_single_reward_judger(judger, group_data_item)
+            all_tasks.extend(tasks)
+            judger_list.append(judger.judger_name)
+            task_len_list.append(task_len_list[-1] + len(tasks))
 
-            flat_tasks_with_names = [task for sample_tasks in tasks_per_sample for task in sample_tasks]
-            results = await asyncio.gather(*flat_tasks_with_names)
+        all_results = await asyncio.gather(*all_tasks)
 
-        import collections.abc
-
-        def flatten(results):
-            for item in results:
-                if isinstance(item, collections.abc.Iterable) and not isinstance(item, (str, bytes, dict)):
-                    yield from item
-                else:
-                    yield item
-
-        flat_results = list(flatten(results))
-        assert len(flat_results) == len(group_data_item) * len(active_judgers), (
-            f"Expected {len(group_data_item) * len(active_judgers)} results, but got {len(flat_results)}"
+        assert len(all_results) == len(group_data_item) * len(active_judgers), (
+            f"Expected {len(group_data_item) * len(active_judgers)} results, but got {len(all_results)}"
         )
-        # 将不同Judger的RLJudgerResponseItem进行组装
+
+        active_judger_results = {}
+        for i in range(active_judgers_len):
+            active_judger_results[judger_list[i]] = all_results[task_len_list[i] : task_len_list[i + 1]]
+
+        # 为每个样本创建一个 RLJudgerResponseItem，不同judger的结果放在同一个item中
         uid_list = [item.uid.observation_id for item in group_data_item]
         judger_response_items_dict = {uid: RLJudgerResponseItem(uid=uid) for uid in uid_list}
-        for result in flat_results:
-            return_uid = result.uid
-            judger_response_items_dict[return_uid].reward.update(result.reward)
-            judger_response_items_dict[return_uid].extra_info.update(result.extra_info)
+        for judger_name, results in active_judger_results.items():
+            for result in results:
+                for data in result:
+                    return_uid = data.uid
+                    judger_response_items_dict[return_uid].reward.update(data.reward)
+                    judger_response_items_dict[return_uid].reward.update({judger_name: data.reward})
+                    judger_response_items_dict[return_uid].extra_info.update(data.extra_info)
         return list(judger_response_items_dict.values())
 
     async def run(
@@ -166,27 +185,25 @@ class JudgerController:
             group_data_item = [group_data_item]
 
         if self.enable_weighted_judgers:
-            # todo: 支持多个judger返回多个value的加权/不加权的逻辑
-            pass
-            # data_source = group_data_item[0].data.data_source
-            # # 如果要使用多个judger并且进行加权打分，则必须在数据集中指定data_source的分数
-            # assert data_source , "No data source found for the given datasets"
-            # judger_names = [judger.judger_name for judger in self.reward_judger]
-            # active_reward_judger = [func for func in self.reward_judger if func.judger_name in data_source]
-            # assert active_reward_judger, (
-            #     f"No active reward judger in {judger_names} found for the given data source {data_source}."
-            # )
-            # judger_response_item = await self._call_custom_reward_judger(active_reward_judger, group_data_item)
-            # judger_return_item = [f"weight_{name}" for name, _ in judger_response_item[0].reward.items()]
-            # for item in judger_response_item:
-            #     final_reward = 0
-            #     for name, weight in data_source.items():
-            #         if name in item.reward:
-            #             final_reward += item.reward[name] * weight
-            #     item.reward["weighted_reward"] = final_reward
+            data_source = group_data_item[0].data.data_source
+            # 如果要使用多个judger并且进行加权打分，则必须在数据集中指定data_source的分数
+            assert data_source, "No data source found for the given datasets when multiple judgers are provided."
+            judger_names = [judger.judger_name for judger in self.reward_judger]
+            active_reward_judger = [func for func in self.reward_judger if func.judger_name in data_source]
+            assert active_reward_judger, (
+                f"No active reward judger in {judger_names} found for the given data source {data_source}."
+            )
+            judger_response_item = await self._call_custom_reward_judger(active_reward_judger, group_data_item)
+
+            # NOTE: 只计算score的加权和
+            for item in judger_response_item:
+                final_reward = 0
+                for name, weight in data_source.items():
+                    if name in item.reward:
+                        final_reward += item.reward[name]["score"] * weight
+                item.reward["weighted_score"] = final_reward
         else:
             judger_response_item = await self._call_custom_reward_judger(self.reward_judger, group_data_item)
-
         if input_type_is_list is False:
             return judger_response_item[0]
         return judger_response_item
